Separate enum AIEventType to fix compilation of GO AI
